Moved group subscription logic to library
